+++ graal-nodejs/mx.graal-nodejs/com.oracle.truffle.trufflenode/src/com/oracle/truffle/trufflenode/serialization/Deserializer.java
@@ -529,7 +529,7 @@ public class Deserializer {
         }
         JSErrorObject error = JSError.create(errorType, realm, message);
         assignId(error);
-        JSObject.set(error, JSError.STACK_NAME, stack);
+        JSError.setFormattedStack(error, stack);
         return error;
     }